move getargspec to traitlets
where it's used
